+++ CHANGELOG.md
@@ -3,11 +3,19 @@ All notable changes to this project will be documented in this file.
 
 ## Next - unreleased
 
-- Add Reactive Extensions.
+- Decouple LINQ interface from Collections.
+- Reactive Extensions initiative.
 - Adding new data structure BinaryTree.
 - Adding new data structure AvlTree.
 - Adding new data structure SplayTree.
 - Adding new data structures LinkedQueue, LinkedStack, LinkedList and LinkedDictionary. 
+
+## 3.2.0 - Unreleased
+### Add
+- Added flatMap method, just like the Scala and Javascript implementation.
+
+### Deprecated
+- Deprecated *slice* method, now we can use the *take*, this is part of the Reactive Extensions initiative.
 
 ## 3.1.2 - 2014-09-08 
 ### Changed